Structure and imports refactor
