+++ src/Routes.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import Header from './components/Header';
+import Header from './components/Header/Header';
 import { APP_NAME, ROUTES } from './constants';
 import AboutMePage from './pages/AboutMePage';
 import ArticlePage from './pages/ArticlePage';
@@ -8,14 +8,16 @@ import HomePage from './pages/HomePage';
 
 const Routes = () => {
     return (
-        <Router>
-            <Header appName={ APP_NAME }/>
-            <Switch>
-                <Route exact path={ ROUTES.HOME } component={ HomePage }/>
-                <Route path={ ROUTES.ABOUT_ME } component={ AboutMePage }/>
-                <Route path={ ROUTES.ARTICLE } component={ ArticlePage }/>
-            </Switch>
-        </Router>
+        <main>
+            <Router>
+                <Header appName={ APP_NAME }/>
+                <Switch>
+                    <Route exact path={ ROUTES.HOME } component={ HomePage }/>
+                    <Route path={ ROUTES.ABOUT_ME } component={ AboutMePage }/>
+                    <Route path={ ROUTES.ARTICLE } component={ ArticlePage }/>
+                </Switch>
+            </Router>
+        </main>
     );
 };
