+++ src/components/Article/Article.css
@@ -1,0 +1,10 @@
+article{
+    border: 1px solid silver;
+}
+article header{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 5rem;
+    background-color: silver;
+}
