+++ src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navigation from './Navigation';
+import Navigation from '../Navigation/Navigation';
 import './Header.css';
 
 interface IProps {
