+++ src/components/Section/Section.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './Section.css';
 
 interface IProps {
     data: string;
