+++ src/index.css
@@ -13,6 +13,12 @@ body {
     font-size: 16px;
 }
 
+main{
+    display: flex;
+    flex-direction: column;
+    margin:2rem;
+}
+
 code {
     font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
